+++ frontend/utils/trpc.ts
@@ -1,5 +1,5 @@
 import { createTRPCReact } from '@trpc/react-query';
-import { AppRouter } from '../../backend';
+import { AppRouter } from '../../backend/server';
 import type { inferRouterOutputs } from '@trpc/server';
 
 export const trpc = createTRPCReact<AppRouter>();
